refactor: better types for promo code tool
